Cast the four size_t narrowings MSVC /W4 named
The Windows job was the one part that could not be measured locally.
Its list is small and specific: six C4267 warnings at four lines, all
size() returning size_t where the public API returns unsigned int -
GameEngine::getBiddingOrder, getPlayerCount and canSeeHand, and
Scoreboard::getRoundCount.

Cast rather than widen the API: seats are unsigned int throughout, and
start() accepts 2 to 6 seats, so the value always fits. The two least
obvious sites say so in a comment.

Note the engine target failed before the test target was compiled, so
this run is also the first sight of MSVC's opinion of the tests.

+++ src/Scoreboard.cpp
@@ -115,7 +115,9 @@ const Round &Scoreboard::getRound(unsigned int index) const
 
 unsigned int Scoreboard::getRoundCount() const
 {
-    return rounds.size();
+    // A schedule is at most a few dozen rounds, so the narrowing is safe; it is
+    // spelled out because size() is a size_t and MSVC says so at /W4.
+    return static_cast<unsigned int>(rounds.size());
 }
 
 unsigned int Scoreboard::getCurrentRoundIndex() const
